realtabbench/requirements.txt
vllm==0.6.3
defog_data==0.1.1
func_timeout==4.3.5
langchain==0.2.5
langchain_core==0.2.7
langchain_experimental==0.0.61
langchain_openai==0.1.8
numpy
pandas
Requests==2.32.3
scikit_learn==1.3.2
sentence_transformers==3.0.1
spacy==3.7.5
sql_metadata==2.11.0
SQLAlchemy==2.0.8
tqdm
matplotlib
mplfonts
seaborn
einops
tabulate==0.9.0
pypinyin
rouge_score==0.1.2
nltk==3.9.1
evaluate==0.4.2
sacrebleu==2.4.3
bert_score==0.3.13
absl-py
einops
babel==2.16.0
openpyxl==3.1.5
datasets==2.21.0